+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/multiple_libs_lib3.sql
@@ -1,0 +1,19 @@
+CREATE TEXT SEARCH TEMPLATE public.test_template (
+    LEXIZE = dsnowball_lexize );
+
+CREATE TEXT SEARCH PARSER public.test_parser (
+    START = prsd_start,
+    GETTOKEN = prsd_nexttoken,
+    END = prsd_end,
+    HEADLINE = prsd_headline,
+    LEXTYPES = prsd_lextype );
+
+CREATE TEXT SEARCH DICTIONARY public.test_dictionary (
+    TEMPLATE = snowball,
+    language = 'english',
+    stopwords = 'english' );
+
+CREATE TEXT SEARCH CONFIGURATION public.test_config (
+    PARSER = pg_catalog."default" );
+
+